requirements.txt
nbconvert>=7.0.0
jupyter>=1.0.0
jinja2>=3.0.0
beautifulsoup4>=4.11.0
pillow>=9.0.0
matplotlib>=3.5.0
pyinstaller>=5.0.0 